+++ src/slurm.rs
@@ -5,6 +5,7 @@ pub mod local {
 
     use crate::{Entity, Modifiable};
 
+    /// TODO: Bubble up error instead of just logging it
     pub fn add_slurm_user(entity: &Entity, sacctmgr_path: &str) {
         let output = Command::new(sacctmgr_path)
             .arg("add")
@@ -41,6 +42,7 @@ pub mod local {
         modify_qos(entity, sacctmgr_path, false);
     }
 
+    /// TODO: Bubble up error instead of just logging it
     pub fn delete_slurm_user(user: &str, sacctmgr_path: &str) {
         let output = Command::new(sacctmgr_path)
             .arg("delete")
@@ -72,6 +74,7 @@ pub mod local {
         }
     }
 
+    /// TODO: Bubble up error instead of just logging it
     pub fn modify_slurm_user(modifiable: &Modifiable, sacctmgr_path: &str) {
         debug!("Start modifying user qos");
         match &modifiable.default_qos {
@@ -108,19 +111,21 @@ pub mod local {
         }
     }
 
+    /// TODO: Bubble up error instead of just logging it
     pub fn list_users(sacctmgr_path: &str) {
         let output = Command::new(sacctmgr_path)
-            .arg("list")
-            .arg("users")
-            .arg("format=User%30,DefaultAccount,Admin%15")
+            .arg("show")
+            .arg("assoc")
+            .arg("format=User%30,Account,DefaultQOS,QOS%80")
             .output()
             .expect(
                 "Unable to execute sacctmgr command. Is the path specified in your config correct?",
             );
-
+        // sacctmgr show assoc format=cluster,account,qos
         println!("{}", String::from_utf8_lossy(&output.stdout));
     }
 
+    /// TODO: Bubble up error instead of just logging it
     fn modify_qos(entity: &Entity, sacctmgr_path: &str, default_qos: bool) {
         let mut qos_str: String = "defaultQos=".to_owned();
         if default_qos {
@@ -129,7 +134,7 @@ pub mod local {
             let qos_joined = entity.qos.join(",");
             qos_str = format!("qos={}", qos_joined);
         }
-
+        debug!("Attempting to modify user with QOS: {qos_str}");
         let output = Command::new(sacctmgr_path)
             .arg("modify")
             .arg("user")
@@ -171,6 +176,7 @@ pub mod remote {
     use crate::ssh::SshCredential;
     use crate::{Entity, Modifiable};
 
+    /// TODO: Bubble up error instead of just logging it
     pub fn add_slurm_user(entity: &Entity, config: &MgmtConfig, credentials: &SshCredential) {
         // Connect to the SSH server and authenticate
         info!("Connecting to {}", config.head_node);
@@ -185,7 +191,7 @@ pub mod remote {
             "{} add user {} Account={} --immediate",
             config.sacctmgr_path, entity.username, entity.group
         );
-        let exit_code = run_command(&sess, &cmd);
+        let exit_code = run_remote_command(&sess, &cmd);
 
         match exit_code {
             0 => info!(
@@ -197,8 +203,13 @@ pub mod remote {
                 cmd
             ),
         };
+
+        debug!("Modifying user qos");
+        modify_qos(&entity, config, &sess, true);
+        modify_qos(&entity, config, &sess, false);
     }
 
+    /// TODO: Bubble up error instead of just logging it
     pub fn delete_slurm_user(user: &str, config: &MgmtConfig, credentials: &SshCredential) {
         // Connect to the SSH server and authenticate
         info!("Connecting to {}", config.head_node);
@@ -210,7 +221,7 @@ pub mod remote {
             .unwrap();
 
         let cmd = format!("{} delete user {} --immediate", config.sacctmgr_path, user);
-        let exit_code = run_command(&sess, &cmd);
+        let exit_code = run_remote_command(&sess, &cmd);
 
         match exit_code {
             0 => info!("Successfully deleted Slurm user {}.", user),
@@ -221,6 +232,7 @@ pub mod remote {
         };
     }
 
+    /// TODO: Bubble up error instead of just logging it
     pub fn modify_slurm_user(
         modifiable: &Modifiable,
         config: &MgmtConfig,
@@ -243,6 +255,7 @@ pub mod remote {
                     default_qos: m.to_string(),
                     ..Default::default()
                 };
+                debug!("New defaultQOS will be {}", entity.default_qos);
                 modify_qos(&entity, config, &sess, true);
             }
             None => info!(
@@ -258,6 +271,7 @@ pub mod remote {
                 qos: modifiable.qos.clone(),
                 ..Default::default()
             };
+            debug!("New QOS will be {:?}", entity.qos);
             modify_qos(&entity, config, &sess, false);
         } else {
             info!(
@@ -267,8 +281,9 @@ pub mod remote {
         }
     }
 
+    /// TODO: Bubble up error instead of just logging it
     pub fn list_users(config: &MgmtConfig, credentials: &SshCredential) {
-        let cmd = "sacctmgr list users format=User%30,DefaultAccount,Admin%15";
+        let cmd = "sacctmgr show assoc format=User%30,Account,DefaultQOS,QOS%80";
 
         // Connect to the SSH server and authenticate
         info!("Connecting to {}", config.head_node);
@@ -287,6 +302,7 @@ pub mod remote {
         println!("{}", s);
     }
 
+    /// TODO: Bubble up error instead of just logging it
     fn modify_qos(entity: &Entity, config: &MgmtConfig, sess: &Session, default_qos: bool) {
         let mut qos_str: String = "defaultQos=".to_owned();
         if default_qos {
@@ -295,12 +311,13 @@ pub mod remote {
             let qos_joined = entity.qos.join(",");
             qos_str = format!("qos={}", qos_joined);
         }
+        debug!("Attempting to modify user QOS with {qos_str}");
 
         let cmd = format!(
             "{} modify user {} set {} --immediate",
             config.sacctmgr_path, entity.username, qos_str
         );
-        let exit_code = run_command(sess, &cmd);
+        let exit_code = run_remote_command(sess, &cmd);
 
         match exit_code {
             0 => info!(
@@ -314,7 +331,11 @@ pub mod remote {
         };
     }
 
-    fn run_command(sess: &Session, cmd: &str) -> i32 {
+    /// Executes given command `cmd` on remote machine over ssh
+    ///
+    /// TODO: move checking if for exit code inside here and return result instead of error code
+    /// directly
+    fn run_remote_command(sess: &Session, cmd: &str) -> i32 {
         debug!("Running command {}", cmd);
 
         let mut channel = sess.channel_session().unwrap();